[DRAFT] New flat config
This is in hiatus because the import module can't ahndle it
https://github.com/import-js/eslint-plugin-import/issues/2556#issuecomment-1678150292
