+++ .eslintignore
@@ -5,8 +5,6 @@ dist/
 .vscode/
 node_modules/
 coverage/
-!.eslintrc.js
-!.prettierrc.js
 !*.js
 !*.ts
 !.vscode/*.json
