+++ .vscode/settings.json
@@ -5,9 +5,16 @@
     ],
     "editor.formatOnSave": true,
     "editor.defaultFormatter": "dbaeumer.vscode-eslint",
-    "[javascript,typescript]": {
-        "editor.defaultFormatter": "dbaeumer.vscode-eslint"
+    "editor.codeActionsOnSave": {
+        "source.organizeImports": "never"
     },
+    "[javascript,typescript]": {
+        "editor.defaultFormatter": "dbaeumer.vscode-eslint",
+        "editor.codeActionsOnSave": {
+            "source.organizeImports": "never"
+        },
+    },
+    "eslint.experimental.useFlatConfig": true,
     "jest.jestCommandLine": "npm test -- ",
     "jest.runMode": {
         "type": "on-demand"
