+++ package.json
@@ -29,6 +29,7 @@
         "internationalization"
     ],
     "devDependencies": {
+        "@eslint/js": "^8.57.0",
         "@stylistic/eslint-plugin": "^1.6.3",
         "@stylistic/eslint-plugin-ts": "^1.6.3",
         "@types/jest": "^29.5.3",
@@ -40,9 +41,11 @@
         "eslint-plugin-import": "^2.28.0",
         "eslint-plugin-jest": "^27.2.3",
         "fast-check": "^3.12.0",
+        "globals": "^14.0.0",
         "jest": "^29.6.2",
         "ts-jest": "^29.1.1",
         "tsc-watch": "^6.0.4",
-        "typescript": "^5.3.2"
+        "typescript": "^5.3.2",
+        "typescript-eslint": "^7.1.1"
     }
 }
